使用decimal包重新设计,但精度还是有问题.

+++ lib/model/geometry/planes/polygon_helper.dart
@@ -176,7 +176,7 @@ extension PolygonPathRelativeWith on Polygon
   }
   Path toPath(){
     Path path = Path();
-    path.addPolygon(points.map((e) => Offset(e.x,e.y)).toList(), true);
+    path.addPolygon(points.map((e) => Offset(e.x.toDouble(),e.y.toDouble())).toList(), true);
     return path;
   }
 }
